@@ -1,5 +1,12 @@
-<script>
-  import '../app.css';
+<script lang="ts">
+	import '../app.css';
+	import favicon from '$lib/assets/favicon.svg';
+	
+	let { children } = $props();
 </script>
 
-<slot />
+<svelte:head>
+	<link rel="icon" href={favicon} />
+</svelte:head>
+
+{@render children?.()}
