@@ -3,10 +3,10 @@ import { vitePreprocess } from '@sveltejs/vite-plugin-svelte';
 
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
+	// Consult https://svelte.dev/docs/kit/integrations
+	// for more information about preprocessors
 	preprocess: vitePreprocess(),
-	kit: {
-		adapter: adapter()
-	}
+	kit: { adapter: adapter() }
 };
 
 export default config;
