@@ -1,6 +1,10 @@
-import { svelte } from '@sveltejs/vite-plugin-svelte';
+import tailwindcss from '@tailwindcss/vite';
+import { sveltekit } from '@sveltejs/kit/vite';
 import { defineConfig } from 'vite';
 
 export default defineConfig({
-	plugins: [svelte()]
+	plugins: [tailwindcss(), sveltekit()],
+	server:{
+		allowedHosts:['5173-09liweis-smsapp-k9jj5wvlehc.ws-us121.gitpod.io']
+	}
 });
